+++ src/services/middlewares/bodyparser.js
@@ -1,0 +1,6 @@
+import bodyParser from 'body-parser';
+
+export default app => app.use(
+    bodyParser.json(),
+    bodyParser.urlencoded({ extended: false })
+);
